refactor: split services into separate files for easier testing
