+++ src/contexts/WalletContext/services/constants.ts
@@ -1,0 +1,4 @@
+export const REFERRAL_AND_SIGNATURES_URL = `${
+  process.env.REACT_APP_REFERRAL_AND_SIGNATURE_SERVICE_URL || 'https://voltz-rest-api.herokuapp.com'
+}`;
+export const UNAVAILABLE_TEXT = 'Service unavailable, please try again shortly';
